WIP: move persistedState.panels and layoutHistory out of Redux state - most reducer tests passing
