+++ app/providers/CurrentLayoutProvider/reducers.test.tsx
@@ -1,0 +1,1239 @@
+/** @jest-environment jsdom */
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+//
+// This file incorporates work covered by the following copyright and
+// permission notice:
+//
+//   Copyright 2018-2021 Cruise LLC
+//
+//   This source code is licensed under the Apache License, Version 2.0,
+//   found at http://www.apache.org/licenses/LICENSE-2.0
+//   You may not use this file except in compliance with the License.
+
+import { renderHook, act } from "@testing-library/react-hooks";
+import { getLeaves, MosaicParent } from "react-mosaic-component";
+
+import { useCurrentLayout } from "@foxglove/studio-base/context/CurrentLayoutContext";
+import { PanelsState } from "@foxglove/studio-base/context/CurrentLayoutContext/actions";
+import CurrentLayoutProvider from "@foxglove/studio-base/providers/CurrentLayoutProvider";
+import {
+  CreateTabPanelPayload,
+  LoadLayoutPayload,
+  MosaicDropTargetPosition,
+} from "@foxglove/studio-base/types/panels";
+import { TAB_PANEL_TYPE } from "@foxglove/studio-base/util/globalConstants";
+import { getPanelTypeFromId } from "@foxglove/studio-base/util/layout";
+
+import { defaultPlaybackConfig } from "./reducers";
+
+function getStore() {
+  const { result } = renderHook(() => useCurrentLayout(), { wrapper: CurrentLayoutProvider });
+
+  return { result };
+}
+
+describe("layout reducers", () => {
+  // FIXME: replacement for deleted localStorage tests?
+
+  describe("adds panel to a layout", () => {
+    it("adds panel to main app layout", () => {
+      const { result } = getStore();
+      const panelLayout = {
+        layout: "Tab!a",
+        configById: {
+          "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A" }, { title: "B" }, { title: "C" }] },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.addPanel({
+          id: "Audio!x",
+          tabId: undefined,
+          layout: panelLayout.layout,
+          config: { foo: "bar" },
+        });
+      });
+      const panels = result.current.state;
+      const layout = panels.layout as MosaicParent<string>;
+      const firstStr = layout.first as string;
+      const secondStr = layout.second as string;
+      expect(layout.direction).toEqual("row");
+      expect(firstStr).toEqual("Audio!x");
+      expect(layout.second).toEqual("Tab!a");
+
+      expect(result.current.state.configById[firstStr]).toEqual({ foo: "bar" });
+      expect(result.current.state.configById[secondStr]).toEqual(panelLayout.configById["Tab!a"]);
+    });
+    it("adds panel to empty Tab layout", () => {
+      const { result } = getStore();
+      const panelLayout = {
+        layout: "Tab!a",
+        configById: {
+          "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A" }, { title: "B" }, { title: "C" }] },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.addPanel({
+          id: "Audio!x",
+          tabId: "Tab!a",
+          layout: undefined,
+          config: { foo: "bar" },
+        });
+      });
+
+      const { layout, configById } = result.current.state;
+      const tabs = configById["Tab!a"]?.tabs;
+      const newAudioId = tabs[0].layout;
+      expect(layout).toEqual("Tab!a");
+      expect(configById["Tab!a"]?.activeTabIdx).toEqual(0);
+      expect(tabs[0].title).toEqual("A");
+      expect(newAudioId).toEqual("Audio!x");
+      expect(tabs.length).toEqual(3);
+
+      expect(configById[newAudioId]).toEqual({ foo: "bar" });
+      expect(configById[layout as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "A", layout: newAudioId }, { title: "B" }, { title: "C" }],
+      });
+    });
+  });
+
+  describe("drops panel into a layout", () => {
+    it("drops panel into app layout", () => {
+      const { result } = getStore();
+      const panelLayout = {
+        layout: "Tab!a",
+        configById: {
+          "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A" }, { title: "B" }, { title: "C" }] },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.dropPanel({
+          newPanelType: "Audio",
+          destinationPath: [],
+          position: "right",
+          config: { foo: "bar" },
+          relatedConfigs: undefined,
+        });
+      });
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(layout.direction).toEqual("row");
+      expect(layout.first).toEqual("Tab!a");
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Audio");
+    });
+    it("drops Tab panel into app layout", () => {
+      const { result } = getStore();
+      const panelLayout = { layout: "Audio!a", configById: { "Audio!a": { foo: "bar" } } };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.dropPanel({
+          newPanelType: "Tab",
+          destinationPath: [],
+          position: "right",
+          config: { activeTabIdx: 0, tabs: [{ title: "A", layout: "Audio!b" }] },
+          relatedConfigs: { "Audio!b": { foo: "baz" } },
+        });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(layout.direction).toEqual("row");
+      expect(layout.first).toEqual("Audio!a");
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Tab");
+
+      expect(configById["Audio!a"]).toEqual({ foo: "bar" });
+      const { activeTabIdx, tabs } = configById[layout.second as string]!;
+      expect(activeTabIdx).toEqual(0);
+      expect(tabs.length).toEqual(1);
+      expect(tabs[0].title).toEqual("A");
+
+      const newAudioId = tabs[0].layout;
+      expect(getPanelTypeFromId(newAudioId)).toEqual("Audio");
+      expect(configById[newAudioId]).toEqual({ foo: "baz" });
+    });
+    it("drops panel into empty Tab layout", () => {
+      const { result } = getStore();
+      const panelLayout = {
+        layout: "Tab!a",
+        configById: {
+          "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A" }, { title: "B" }, { title: "C" }] },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.dropPanel({
+          newPanelType: "Audio",
+          destinationPath: [],
+          position: "right",
+          tabId: "Tab!a",
+          config: { foo: "bar" },
+          relatedConfigs: undefined,
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual("Tab!a");
+      const tabs = configById["Tab!a"]?.tabs;
+      expect(tabs[0].title).toEqual("A");
+      expect(getPanelTypeFromId(tabs[0].layout)).toEqual("Audio");
+      expect(tabs.length).toEqual(3);
+    });
+    it("drops panel into nested Tab layout", () => {
+      const { result } = getStore();
+      const panelLayout = {
+        layout: "Tab!a",
+        configById: {
+          "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A", layout: "Tab!b" }] },
+          "Tab!b": { activeTabIdx: 0, tabs: [{ title: "B", layout: "Plot!a" }] },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.dropPanel({
+          newPanelType: "Audio",
+          destinationPath: [],
+          position: "right",
+          tabId: "Tab!b",
+          config: { foo: "bar" },
+          relatedConfigs: undefined,
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual("Tab!a");
+      expect(configById["Tab!a"]).toEqual(panelLayout.configById["Tab!a"]);
+      const tabBTabs = configById["Tab!b"]?.tabs;
+      expect(tabBTabs.length).toEqual(1);
+      expect(tabBTabs[0].layout.first).toEqual("Plot!a");
+      expect(getPanelTypeFromId(tabBTabs[0].layout.second)).toEqual("Audio");
+      expect(configById[tabBTabs[0].layout.second]).toEqual({ foo: "bar" });
+    });
+    it("drops nested Tab panel into main layout", () => {
+      const { result } = getStore();
+      const panelLayout = { layout: "Audio!a" };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.dropPanel({
+          newPanelType: "Tab",
+          destinationPath: [],
+          position: "right",
+          config: { activeTabIdx: 0, tabs: [{ title: "A", layout: "Tab!b" }] },
+          relatedConfigs: {
+            "Tab!b": { activeTabIdx: 0, tabs: [{ title: "B", layout: "Plot!a" }] },
+          },
+        });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(layout.first).toEqual("Audio!a");
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Tab");
+
+      const parentTabConfig = configById[layout.second as string];
+      expect(parentTabConfig?.tabs.length).toEqual(1);
+      expect(parentTabConfig?.tabs[0].title).toEqual("A");
+
+      const childTabId = parentTabConfig?.tabs[0].layout;
+      expect(getPanelTypeFromId(childTabId)).toEqual("Tab");
+      const childTabProps = configById[childTabId];
+      expect(childTabProps?.activeTabIdx).toEqual(0);
+      expect(childTabProps?.tabs.length).toEqual(1);
+      expect(childTabProps?.tabs[0].title).toEqual("B");
+      expect(getPanelTypeFromId(childTabProps?.tabs[0].layout)).toEqual("Plot");
+    });
+  });
+
+  describe("moves tabs", () => {
+    it("reorders tabs within a Tab panel", () => {
+      const { result } = getStore();
+      const panelLayout = {
+        layout: "Tab!a",
+        configById: {
+          "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A" }, { title: "B" }, { title: "C" }] },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.moveTab({
+          source: { panelId: "Tab!a", tabIndex: 0 },
+          target: { panelId: "Tab!a", tabIndex: 1 },
+        });
+      });
+      const { configById } = result.current.state;
+      expect(configById).toEqual({
+        "Tab!a": { activeTabIdx: 1, tabs: [{ title: "B" }, { title: "A" }, { title: "C" }] },
+      });
+    });
+
+    it("moves tabs between Tab panels", () => {
+      const { result } = getStore();
+      const layout: MosaicParent<string> = { first: "Tab!a", second: "Tab!b", direction: "row" };
+      const panelLayout = {
+        layout,
+        configById: {
+          "Tab!a": {
+            activeTabIdx: 0,
+            tabs: [{ title: "A" }, { title: "B" }, { title: "C" }],
+          },
+          "Tab!b": {
+            activeTabIdx: 0,
+            tabs: [{ title: "D" }, { title: "E" }, { title: "F" }],
+          },
+        },
+      };
+      act(() => {
+        result.current.loadLayout(panelLayout);
+        result.current.moveTab({
+          source: { panelId: "Tab!a", tabIndex: 0 },
+          target: { panelId: "Tab!b", tabIndex: 1 },
+        });
+      });
+      const { configById } = result.current.state;
+      expect(configById).toEqual({
+        "Tab!a": { activeTabIdx: 0, tabs: [{ title: "B" }, { title: "C" }] },
+        "Tab!b": {
+          activeTabIdx: 0,
+          tabs: [{ title: "D" }, { title: "A" }, { title: "E" }, { title: "F" }],
+        },
+      });
+    });
+  });
+
+  it.skip("closes a panel in single-panel layout", () => {
+    const { result } = getStore();
+    act(() => {
+      result.current.loadLayout({ layout: "Audio!a", configById: { "Audio!a": { foo: "bar" } } });
+      result.current.setSelectedPanelIds(["Audio!a", "unknown!b"]);
+      result.current.closePanel({ root: "Audio!a", path: [] });
+    });
+    checkState(({ mosaic: { selectedPanelIds } }) => {
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual(undefined);
+      expect(configById).toEqual({});
+      expect(selectedPanelIds).toEqual(["unknown!b"]);
+    });
+  });
+
+  it.skip("closes a panel in multi-panel layout", () => {
+    const { result } = getStore();
+    const layout: MosaicParent<string> = { first: "Audio!a", second: "Audio!b", direction: "row" };
+    const panelLayout = {
+      layout,
+      configById: { "Audio!a": { foo: "bar" }, "Audio!b": { foo: "baz" } },
+    };
+    act(() => {
+      result.current.loadLayout(panelLayout);
+      result.current.setSelectedPanelIds(["Audio!a", "Audio!b"]);
+      result.current.closePanel({ root: panelLayout.layout, path: ["first"] });
+    });
+    checkState(({ mosaic: { selectedPanelIds } }) => {
+      expect(result.current.state.layout).toEqual("Audio!b");
+      expect(result.current.state.configById).toEqual({ "Audio!b": { foo: "baz" } });
+      expect(selectedPanelIds).toEqual(["Audio!b"]);
+    });
+  });
+
+  it.skip("closes a panel nested inside a Tab panel", () => {
+    const { result } = getStore();
+    const panelLayout = {
+      layout: "Tab!a",
+      configById: {
+        "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A", layout: "Audio!a" }] },
+        "Audio!a": { foo: "bar" },
+      },
+    };
+    act(() => {
+      result.current.loadLayout(panelLayout);
+      result.current.setSelectedPanelIds(["Audio!a"]);
+      result.current.closePanel({ root: "Audio!a", path: [], tabId: "Tab!a" });
+    });
+    checkState(({ mosaic: { selectedPanelIds } }) => {
+      expect(result.current.state.layout).toEqual("Tab!a");
+      expect(result.current.state.configById).toEqual({
+        "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A", layout: undefined }] },
+      });
+      expect(selectedPanelIds).toEqual([]);
+    });
+  });
+
+  it("loads a layout", () => {
+    const { result } = getStore();
+    const panelsState = { layout: { foo: "baz" } };
+    act(() => {
+      result.current.loadLayout(panelsState);
+    });
+    expect(result.current.state.layout).toEqual({ foo: "baz" });
+  });
+
+  it("resets panels to a valid state when importing an empty layout", () => {
+    const { result } = getStore();
+    act(() => {
+      result.current.loadLayout({ layout: undefined });
+    });
+    expect(result.current.state).toEqual({
+      globalVariables: {},
+      layout: undefined,
+      linkedGlobalVariables: [],
+      playbackConfig: defaultPlaybackConfig,
+      configById: {},
+      userNodes: {},
+    });
+  });
+
+  it.skip("will set local storage when importing a panel layout", () => {
+    const { result } = getStore();
+
+    act(() => {
+      result.current.loadLayout({ layout: "myNewLayout", configById: {} });
+    });
+    checkState(() => {
+      const globalState = GetGlobalState();
+      expect(globalState.panels.layout).toEqual("myNewLayout");
+    });
+  });
+
+  describe("creates Tab panels from existing panels correctly", () => {
+    const regularLayoutPayload = {
+      layout: {
+        first: "Audio!a",
+        second: { first: "RawMessages!a", second: "Audio!c", direction: "column" },
+        direction: "row",
+      },
+      configById: { "Audio!a": { foo: "bar" }, "RawMessages!a": { foo: "baz" } },
+    } as LoadLayoutPayload;
+    const basePayload = {
+      idToReplace: "Audio!a",
+      newId: "Tab!a",
+      idsToRemove: ["Audio!a", "RawMessages!a"],
+      singleTab: false,
+    };
+    const nestedLayoutPayload = {
+      layout: {
+        first: "Audio!a",
+        second: "Tab!z",
+        direction: "column",
+      },
+      configById: {
+        "Audio!a": { foo: "bar" },
+        "Tab!z": {
+          activeTabIdx: 0,
+          tabs: [
+            {
+              title: "First tab",
+              layout: { first: "Audio!b", second: "RawMessages!a", direction: "row" },
+            },
+          ],
+        },
+        "Audio!b": { foo: "baz" },
+        "RawMessages!a": { raw: "messages" },
+      },
+    } as LoadLayoutPayload;
+    const createTabPanelPayload = {
+      ...basePayload,
+      layout: regularLayoutPayload.layout,
+    } as CreateTabPanelPayload;
+    const nestedCreateTabPanelPayload = {
+      ...basePayload,
+      layout: nestedLayoutPayload.layout,
+    } as CreateTabPanelPayload;
+
+    it("will group selected panels into a Tab panel", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout(regularLayoutPayload);
+        result.current.createTabPanel({ ...createTabPanelPayload, singleTab: true });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(getPanelTypeFromId(layout.first as string)).toEqual(TAB_PANEL_TYPE);
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Audio");
+      expect(configById[layout.first as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          {
+            title: "1",
+            layout: { direction: "row", first: "Audio!a", second: "RawMessages!a" },
+          },
+        ],
+      });
+      expect(configById[layout.second as string]).toEqual(
+        regularLayoutPayload.configById?.[layout.second as string],
+      );
+    });
+
+    it("will group selected panels into a Tab panel, even when a selected panel is nested", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout(nestedLayoutPayload);
+        result.current.createTabPanel({ ...nestedCreateTabPanelPayload, singleTab: true });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(getPanelTypeFromId(layout.first as string)).toEqual(TAB_PANEL_TYPE);
+      expect(getPanelTypeFromId(layout.second as string)).toEqual(TAB_PANEL_TYPE);
+      expect(configById[layout.first as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          {
+            title: "1",
+            layout: { direction: "column", first: "Audio!a", second: "RawMessages!a" },
+          },
+        ],
+      });
+      expect(configById[layout.second as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "First tab", layout: "Audio!b" }],
+      });
+    });
+
+    it("will create individual tabs for selected panels in a new Tab panel", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout(regularLayoutPayload);
+        result.current.createTabPanel({ ...createTabPanelPayload, singleTab: false });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(getPanelTypeFromId(layout.first as string)).toEqual(TAB_PANEL_TYPE);
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Audio");
+      expect(configById[layout.first as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          { title: "Audio", layout: "Audio!a" },
+          { title: "RawMessages", layout: "RawMessages!a" },
+        ],
+      });
+      expect(configById[layout.second as string]).toEqual(
+        regularLayoutPayload.configById?.[layout.second as string],
+      );
+    });
+
+    it("will create individual tabs for selected panels in a new Tab panel, even when a selected panel is nested", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout(nestedLayoutPayload);
+        result.current.createTabPanel({ ...nestedCreateTabPanelPayload, singleTab: false });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(getPanelTypeFromId(layout.first as string)).toEqual(TAB_PANEL_TYPE);
+      expect(getPanelTypeFromId(layout.second as string)).toEqual(TAB_PANEL_TYPE);
+      expect(configById[layout.first as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          { title: "Audio", layout: "Audio!a" },
+          { title: "RawMessages", layout: "RawMessages!a" },
+        ],
+      });
+      expect(configById[layout.second as string]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "First tab", layout: "Audio!b" }],
+      });
+    });
+  });
+
+  it("saves and overwrites user nodes", () => {
+    const { result } = getStore();
+    const firstPayload = { foo: { name: "foo", sourceCode: "bar" } };
+    const secondPayload = { bar: { name: "bar", sourceCode: "baz" } };
+
+    act(() => {
+      result.current.setUserNodes(firstPayload);
+    });
+    expect(result.current.state.userNodes).toEqual(firstPayload);
+
+    act(() => {
+      result.current.setUserNodes(secondPayload);
+    });
+    expect(result.current.state.userNodes).toEqual({ ...firstPayload, ...secondPayload });
+  });
+
+  describe("panel toolbar actions", () => {
+    it("can split panel", () => {
+      const { result } = getStore();
+      const audioConfig = { foo: "bar" };
+      act(() => {
+        result.current.changePanelLayout({ layout: "Audio!a" });
+
+        result.current.savePanelConfigs({ configs: [{ id: "Audio!a", config: audioConfig }] });
+
+        result.current.splitPanel({
+          id: "Audio!a",
+          config: audioConfig,
+          direction: "row",
+          path: [],
+          root: "Audio!a",
+        });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(layout.first).toEqual("Audio!a");
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Audio");
+      expect(layout.direction).toEqual("row");
+      expect(configById["Audio!a"]).toEqual(audioConfig);
+      expect(configById[layout.second as string]).toEqual(audioConfig);
+    });
+
+    it("can split Tab panel", () => {
+      const { result } = getStore();
+      const audioConfig = { foo: "bar" };
+      const tabConfig = { activeTabIdx: 0, tabs: [{ title: "A", layout: "Audio!a" }] };
+      act(() => {
+        result.current.changePanelLayout({ layout: "Tab!a" });
+
+        result.current.savePanelConfigs({
+          configs: [
+            { id: "Tab!a", config: tabConfig },
+            { id: "Audio!a", config: audioConfig },
+          ],
+        });
+
+        result.current.splitPanel({
+          id: "Tab!a",
+          config: tabConfig,
+          direction: "row",
+          path: [],
+          root: "Tab!a",
+        });
+      });
+
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as MosaicParent<string>;
+      expect(layout.first).toEqual("Tab!a");
+      expect(getPanelTypeFromId(layout.second as string)).toEqual("Tab");
+      expect(layout.direction).toEqual("row");
+      expect(configById["Tab!a"]).toEqual(tabConfig);
+      expect(getPanelTypeFromId(configById[layout.second as string]?.tabs[0].layout)).toEqual(
+        "Audio",
+      );
+      expect(configById["Audio!a"]).toEqual(audioConfig);
+    });
+
+    it("can split panel inside Tab panel", () => {
+      const { result } = getStore();
+      const audioConfig = { foo: "bar" };
+      const tabConfig = { activeTabIdx: 0, tabs: [{ title: "A", layout: "Audio!a" }] };
+      act(() => {
+        result.current.changePanelLayout({ layout: "Tab!a" });
+
+        result.current.savePanelConfigs({
+          configs: [
+            { id: "Tab!a", config: tabConfig },
+            { id: "Audio!a", config: audioConfig },
+          ],
+        });
+
+        result.current.splitPanel({
+          id: "Audio!a",
+          tabId: "Tab!a",
+          config: audioConfig,
+          direction: "row",
+          path: [],
+          root: "Audio!a",
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual("Tab!a");
+      const tabLayout = configById["Tab!a"]?.tabs[0].layout;
+      expect(tabLayout.first).toEqual("Audio!a");
+      expect(getPanelTypeFromId(tabLayout.second)).toEqual("Audio");
+      expect(tabLayout.direction).toEqual("row");
+      expect(configById["Audio!a"]).toEqual(audioConfig);
+      expect(configById[tabLayout.second]).toEqual(audioConfig);
+    });
+
+    it("can swap panels", () => {
+      const { result } = getStore();
+      const audioConfig = { foo: "bar" };
+      const rawMessagesConfig = { foo: "baz" };
+      act(() => {
+        result.current.changePanelLayout({ layout: "Audio!a" });
+        result.current.savePanelConfigs({ configs: [{ id: "Audio!a", config: audioConfig }] });
+
+        result.current.swapPanel({
+          originalId: "Audio!a",
+          type: "RawMessages",
+          config: rawMessagesConfig,
+          path: [],
+          root: "Audio!a",
+        });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as string;
+      expect(getPanelTypeFromId(layout)).toEqual("RawMessages");
+      expect(configById["Audio!a"]).toEqual(undefined);
+      expect(configById[layout]).toEqual(rawMessagesConfig);
+    });
+
+    it("can swap panel for a Tab panel", () => {
+      const { result } = getStore();
+      const audioConfig = { foo: "bar" };
+      const tabConfig = { activeTabIdx: 0, tabs: [{ title: "A", layout: "RawMessages!a" }] };
+      const rawMessagesConfig = { path: "foo" };
+      act(() => {
+        result.current.changePanelLayout({ layout: "Audio!a" });
+
+        result.current.savePanelConfigs({ configs: [{ id: "Audio!a", config: audioConfig }] });
+
+        result.current.swapPanel({
+          originalId: "Audio!a",
+          type: "Tab",
+          config: tabConfig,
+          relatedConfigs: { "RawMessages!a": rawMessagesConfig },
+          path: [],
+          root: "Audio!a",
+        });
+      });
+      const { configById } = result.current.state;
+      const layout = result.current.state.layout as string;
+      expect(getPanelTypeFromId(layout)).toEqual("Tab");
+      const tabLayout = configById[layout]?.tabs[0].layout;
+      expect(getPanelTypeFromId(tabLayout)).toEqual("RawMessages");
+      expect(configById[tabLayout]).toEqual(rawMessagesConfig);
+    });
+
+    it("can swap panel inside a Tab", () => {
+      const { result } = getStore();
+      const rawMessagesConfig = { foo: "baz" };
+      const tabConfig = { activeTabIdx: 0, tabs: [{ title: "A", layout: "Audio!a" }] };
+
+      act(() => {
+        result.current.changePanelLayout({ layout: "Tab!a" });
+
+        result.current.savePanelConfigs({
+          configs: [{ id: "Tab!a", config: tabConfig }],
+        });
+        result.current.swapPanel({
+          originalId: "Audio!a",
+          tabId: "Tab!a",
+          type: "RawMessages",
+          config: rawMessagesConfig,
+          path: [],
+          root: "Audio!a",
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual("Tab!a");
+      const tabLayout = configById["Tab!a"]?.tabs[0].layout;
+      expect(getPanelTypeFromId(tabLayout)).toEqual("RawMessages");
+      expect(configById[tabLayout]).toEqual(rawMessagesConfig);
+    });
+  });
+
+  describe("clearing old saved config", () => {
+    const panelState = {
+      layout: {
+        direction: "row",
+        first: "FirstPanel!34otwwt",
+        second: {
+          direction: "column",
+          second: "SecondPanel!2wydzut",
+          first: { direction: "row", second: "ThirdPanel!ye6m1m", first: "FourthPanel!abc" },
+        },
+      },
+      configById: {},
+    } as PanelsState;
+    const tabPanelState = {
+      layout: {
+        direction: "row",
+        first: "FirstPanel!34otwwt",
+        second: {
+          direction: "column",
+          second: "SecondPanel!2wydzut",
+          first: { direction: "row", second: "ThirdPanel!ye6m1m", first: "Tab!abc" },
+        },
+      },
+      configById: {},
+    } as PanelsState;
+
+    it("removes a panel's configById when it is removed from the layout", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.changePanelLayout({ layout: panelState.layout });
+      });
+      // eslint-disable-next-line no-restricted-syntax
+      const leaves = getLeaves(panelState.layout ?? null);
+      expect(leaves).toHaveLength(4);
+      expect(leaves).toContain("FirstPanel!34otwwt");
+      expect(leaves).toContain("SecondPanel!2wydzut");
+      expect(leaves).toContain("ThirdPanel!ye6m1m");
+      expect(leaves).toContain("FourthPanel!abc");
+      expect(result.current.state.configById).toEqual({});
+
+      const panelConfig = {
+        id: "SecondPanel!2wydzut",
+        config: { foo: "bar" },
+        defaultConfig: { foo: "" },
+      };
+      act(() => {
+        result.current.savePanelConfigs({
+          configs: [
+            panelConfig,
+            { id: "FirstPanel!34otwwt", config: { baz: true }, defaultConfig: { baz: false } },
+          ],
+        });
+      });
+      expect(result.current.state.configById).toEqual({
+        "SecondPanel!2wydzut": { foo: "bar" },
+        "FirstPanel!34otwwt": { baz: true },
+      });
+      act(() => {
+        result.current.changePanelLayout({
+          layout: { direction: "row", first: "FirstPanel!34otwwt", second: "SecondPanel!2wydzut" },
+        });
+      });
+      expect(result.current.state.configById).toEqual({
+        "SecondPanel!2wydzut": { foo: "bar" },
+        "FirstPanel!34otwwt": { baz: true },
+      });
+      act(() => {
+        result.current.changePanelLayout({
+          layout: { direction: "row", first: "FirstPanel!34otwwt", second: "ThirdPanel!ye6m1m" },
+        });
+      });
+      expect(result.current.state.configById).toEqual({
+        "FirstPanel!34otwwt": { baz: true },
+      });
+      act(() => {
+        result.current.changePanelLayout({ layout: "foo!1234" });
+      });
+      expect(result.current.state.configById).toEqual({});
+      act(() => {
+        result.current.savePanelConfigs({
+          configs: [{ id: "foo!1234", config: { okay: true }, defaultConfig: { okay: false } }],
+        });
+      });
+      expect(result.current.state.configById).toEqual({
+        "foo!1234": { okay: true },
+      });
+    });
+
+    it("removes a panel's configById when it is removed from Tab panel", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.changePanelLayout({ layout: tabPanelState.layout });
+      });
+      // eslint-disable-next-line no-restricted-syntax
+      const leaves = getLeaves(tabPanelState.layout ?? null);
+      expect(leaves).toHaveLength(4);
+      expect(leaves).toContain("FirstPanel!34otwwt");
+      expect(leaves).toContain("SecondPanel!2wydzut");
+      expect(leaves).toContain("ThirdPanel!ye6m1m");
+      expect(leaves).toContain("Tab!abc");
+      expect(result.current.state.configById).toEqual({});
+
+      const baseTabConfig = {
+        id: "Tab!abc",
+        config: { tabs: [{ title: "Tab A", layout: "NestedPanel!xyz" }], activeTabIdx: 0 },
+      };
+      const nestedPanelConfig = { id: "NestedPanel!xyz", config: { foo: "bar" } };
+      act(() => {
+        result.current.savePanelConfigs({ configs: [baseTabConfig] });
+      });
+      expect(result.current.state.configById).toEqual({ "Tab!abc": baseTabConfig.config });
+
+      act(() => {
+        result.current.savePanelConfigs({ configs: [nestedPanelConfig] });
+      });
+      expect(result.current.state.configById).toEqual({
+        "Tab!abc": baseTabConfig.config,
+        "NestedPanel!xyz": nestedPanelConfig.config,
+      });
+
+      const emptyTabConfig = {
+        id: "Tab!abc",
+        config: { tabs: [{ title: "Tab A", layout: undefined }], activeTabIdx: 0 },
+      };
+      act(() => {
+        result.current.savePanelConfigs({ configs: [emptyTabConfig] });
+      });
+      // "NestedPanel!xyz" key in configById should be gone
+      expect(result.current.state.configById).toEqual({ "Tab!abc": emptyTabConfig.config });
+    });
+
+    it("does not remove old configById when trimConfigById = false", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.changePanelLayout({ layout: "foo!bar" });
+        result.current.savePanelConfigs({ configs: [{ id: "foo!bar", config: { foo: "baz" } }] });
+        result.current.changePanelLayout({ layout: tabPanelState.layout });
+      });
+      expect(result.current.state.configById).toEqual({});
+
+      act(() => {
+        result.current.changePanelLayout({ layout: "foo!bar" });
+        result.current.savePanelConfigs({ configs: [{ id: "foo!bar", config: { foo: "baz" } }] });
+        result.current.changePanelLayout({ layout: tabPanelState.layout, trimConfigById: false });
+      });
+      expect(result.current.state.configById).toEqual({ "foo!bar": { foo: "baz" } });
+    });
+  });
+
+  describe("handles dragging panels", () => {
+    it("does not remove panel from single-panel layout when starting drag", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout({ layout: "Audio!a", configById: {} });
+        result.current.startDrag({ sourceTabId: undefined, path: [] });
+      });
+      expect(result.current.state.layout).toEqual("Audio!a");
+    });
+    it("hides panel from multi-panel layout when starting drag", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout({
+          layout: { first: "Audio!a", second: "RawMessages!a", direction: "column" },
+          configById: {},
+        });
+        result.current.startDrag({ sourceTabId: undefined, path: ["second"] });
+      });
+      expect(result.current.state.layout).toEqual({
+        first: "Audio!a",
+        second: "RawMessages!a",
+        direction: "column",
+        splitPercentage: 100,
+      });
+    });
+    it("removes non-Tab panel from single-panel tab layout when starting drag", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout({
+          layout: { first: "Tab!a", second: "RawMessages!a", direction: "column" },
+          configById: {
+            "Tab!a": { activeTabIdx: 0, tabs: [{ title: "A", layout: "Audio!a" }] },
+          },
+        });
+        result.current.startDrag({ sourceTabId: "Tab!a", path: [] });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual({ first: "Tab!a", second: "RawMessages!a", direction: "column" });
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "A", layout: undefined }],
+      });
+    });
+    it("hides panel from multi-panel Tab layout when starting drag", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout({
+          layout: { first: "Tab!a", second: "RawMessages!a", direction: "column" },
+          configById: {
+            "Tab!a": {
+              activeTabIdx: 0,
+              tabs: [
+                { title: "A", layout: { first: "Audio!a", second: "Plot!a", direction: "row" } },
+              ],
+            },
+          },
+        });
+        result.current.startDrag({ sourceTabId: "Tab!a", path: ["first"] });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual({ first: "Tab!a", second: "RawMessages!a", direction: "column" });
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          {
+            title: "A",
+            layout: {
+              first: "Audio!a",
+              second: "Plot!a",
+              direction: "row",
+              splitPercentage: 0,
+            },
+          },
+        ],
+      });
+    });
+    it("handles drags within the same tab", () => {
+      const { result } = getStore();
+      const originalSavedProps = {
+        "Tab!a": {
+          activeTabIdx: 0,
+          tabs: [{ title: "A", layout: { first: "Audio!a", second: "Plot!a", direction: "row" } }],
+        },
+      };
+      act(() => {
+        result.current.loadLayout({ layout: "Tab!a", configById: originalSavedProps });
+        result.current.startDrag({ sourceTabId: "Tab!a", path: ["first"] });
+        result.current.endDrag({
+          originalLayout: "Tab!a",
+          originalSavedProps,
+          panelId: "Audio!a",
+          sourceTabId: "Tab!a",
+          targetTabId: "Tab!a",
+          position: "right",
+          destinationPath: ["second"],
+          ownPath: ["first"],
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual("Tab!a");
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "A", layout: { first: "Plot!a", second: "Audio!a", direction: "row" } }],
+      });
+    });
+    it("handles drags to main layout from Tab panel", () => {
+      const { result } = getStore();
+      const originalLayout = {
+        first: "Tab!a",
+        second: "RawMessages!a",
+        direction: "column",
+      } as MosaicParent<string>;
+      const originalSavedProps = {
+        "Tab!a": {
+          activeTabIdx: 0,
+          tabs: [{ title: "A", layout: { first: "Audio!a", second: "Plot!a", direction: "row" } }],
+        },
+      };
+      act(() => {
+        result.current.loadLayout({ layout: originalLayout, configById: originalSavedProps });
+        result.current.startDrag({ sourceTabId: "Tab!a", path: ["first"] });
+        result.current.endDrag({
+          originalLayout,
+          originalSavedProps,
+          panelId: "Audio!a",
+          sourceTabId: "Tab!a",
+          targetTabId: undefined,
+          position: "right",
+          destinationPath: ["second"],
+          ownPath: ["first"],
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual({
+        first: "Tab!a",
+        second: { first: "RawMessages!a", second: "Audio!a", direction: "row" },
+        direction: "column",
+      });
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "A", layout: "Plot!a" }],
+      });
+    });
+    it("handles drags to Tab panel from main layout", () => {
+      const { result } = getStore();
+      const originalLayout = {
+        first: "Tab!a",
+        second: "RawMessages!a",
+        direction: "column",
+      } as MosaicParent<string>;
+      const originalSavedProps = {
+        "Tab!a": {
+          activeTabIdx: 0,
+          tabs: [{ title: "A", layout: { first: "Audio!a", second: "Plot!a", direction: "row" } }],
+        },
+      };
+      act(() => {
+        result.current.loadLayout({
+          layout: originalLayout,
+          configById: originalSavedProps,
+        });
+        result.current.startDrag({ sourceTabId: undefined, path: ["second"] });
+        result.current.endDrag({
+          originalLayout,
+          originalSavedProps,
+          panelId: "RawMessages!a",
+          sourceTabId: undefined,
+          targetTabId: "Tab!a",
+          position: "right",
+          destinationPath: ["second"],
+          ownPath: ["second"],
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual("Tab!a");
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          {
+            title: "A",
+            layout: {
+              first: "Audio!a",
+              second: { first: "Plot!a", second: "RawMessages!a", direction: "row" },
+              direction: "row",
+            },
+          },
+        ],
+      });
+    });
+    it("handles dragging non-Tab panels between Tab panels", () => {
+      const { result } = getStore();
+      const originalLayout = {
+        first: "Tab!a",
+        second: "Tab!b",
+        direction: "column",
+      } as MosaicParent<string>;
+      const originalSavedProps = {
+        "Tab!a": {
+          activeTabIdx: 0,
+          tabs: [{ title: "A", layout: { first: "Audio!a", second: "Plot!a", direction: "row" } }],
+        },
+        "Tab!b": {
+          activeTabIdx: 0,
+          tabs: [
+            {
+              title: "B",
+              layout: { first: "RawMessages!a", second: "Publish!a", direction: "row" },
+            },
+          ],
+        },
+      };
+      act(() => {
+        result.current.loadLayout({ layout: originalLayout, configById: originalSavedProps });
+        result.current.startDrag({ sourceTabId: "Tab!a", path: ["first"] });
+        result.current.endDrag({
+          originalLayout,
+          originalSavedProps,
+          panelId: "Audio!a",
+          sourceTabId: "Tab!a",
+          targetTabId: "Tab!b",
+          position: "right",
+          destinationPath: ["first"],
+          ownPath: ["first"],
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual({ first: "Tab!a", second: "Tab!b", direction: "column" });
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "A", layout: "Plot!a" }],
+      });
+      expect(configById["Tab!b"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [
+          {
+            title: "B",
+            layout: {
+              first: { first: "RawMessages!a", second: "Audio!a", direction: "row" },
+              second: "Publish!a",
+              direction: "row",
+            },
+          },
+        ],
+      });
+    });
+    it("handles dragging Tab panels between Tab panels", () => {
+      const { result } = getStore();
+      const originalLayout = {
+        first: "Tab!a",
+        second: "Tab!b",
+        direction: "column",
+      } as MosaicParent<string>;
+      const tabCConfig = {
+        activeTabIdx: 0,
+        tabs: [
+          { title: "C1", layout: "Plot!a" },
+          { title: "C2", layout: undefined },
+        ],
+      };
+      const originalSavedProps = {
+        "Tab!a": {
+          activeTabIdx: 0,
+          tabs: [{ title: "A", layout: { first: "Tab!c", second: "Audio!a", direction: "row" } }],
+        },
+        "Tab!b": { activeTabIdx: 0, tabs: [{ title: "B" }] },
+        "Tab!c": tabCConfig,
+      };
+      act(() => {
+        result.current.loadLayout({ layout: originalLayout, configById: originalSavedProps });
+        result.current.startDrag({ sourceTabId: "Tab!a", path: ["first"] });
+        result.current.endDrag({
+          originalLayout,
+          originalSavedProps,
+          panelId: "Tab!c",
+          sourceTabId: "Tab!a",
+          targetTabId: "Tab!b",
+          position: "right",
+          destinationPath: [],
+          ownPath: ["first"],
+        });
+      });
+      const { layout, configById } = result.current.state;
+      expect(layout).toEqual({ first: "Tab!a", second: "Tab!b", direction: "column" });
+      expect(configById["Tab!a"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "A", layout: "Audio!a" }],
+      });
+      expect(configById["Tab!b"]).toEqual({
+        activeTabIdx: 0,
+        tabs: [{ title: "B", layout: "Tab!c" }],
+      });
+      expect(configById["Tab!c"]).toEqual(tabCConfig);
+    });
+    it("handles drags in single-panel layouts", () => {
+      const { result } = getStore();
+      act(() => {
+        result.current.loadLayout({ layout: "Audio!a" });
+        result.current.startDrag({ sourceTabId: undefined, path: [] });
+        result.current.endDrag({
+          originalLayout: "Audio!a",
+          originalSavedProps: {},
+          panelId: "Audio!a",
+          sourceTabId: undefined,
+          targetTabId: undefined,
+          position: "right",
+          destinationPath: [],
+          ownPath: [],
+        });
+      });
+      expect(result.current.state.layout).toEqual("Audio!a");
+    });
+    it("handles drags in multi-panel layouts", () => {
+      const { result } = getStore();
+      const originalLayout = {
+        first: "Audio!a",
+        second: "Plot!a",
+        direction: "row",
+      } as MosaicParent<string>;
+      act(() => {
+        result.current.loadLayout({ layout: originalLayout });
+        result.current.startDrag({ sourceTabId: undefined, path: ["first"] });
+        result.current.endDrag({
+          originalLayout,
+          originalSavedProps: {},
+          panelId: "Audio!a",
+          sourceTabId: undefined,
+          targetTabId: undefined,
+          position: "right",
+          destinationPath: ["second"],
+          ownPath: ["first"],
+        });
+      });
+      expect(result.current.state.layout).toEqual({
+        first: "Plot!a",
+        second: "Audio!a",
+        direction: "row",
+      });
+    });
+    it("handles drags in multi-panel layouts - invalid position", () => {
+      const { result } = getStore();
+      const originalLayout = {
+        first: "Audio!a",
+        second: "Plot!a",
+        direction: "row",
+      } as MosaicParent<string>;
+      act(() => {
+        result.current.loadLayout({ layout: originalLayout });
+        result.current.startDrag({ sourceTabId: undefined, path: ["first"] });
+        result.current.endDrag({
+          originalLayout,
+          originalSavedProps: {},
+          panelId: "Audio!a",
+          sourceTabId: undefined,
+          targetTabId: undefined,
+          position: undefined as unknown as MosaicDropTargetPosition,
+          destinationPath: ["second"],
+          ownPath: ["first"],
+        });
+      });
+      expect(result.current.state.layout).toEqual({
+        first: "Audio!a",
+        second: "Plot!a",
+        direction: "row",
+        splitPercentage: undefined,
+      });
+    });
+  });
+});
