+++ app/reducers/index.ts
@@ -23,17 +23,11 @@ import layoutHistory, {
   initialLayoutHistoryState,
 } from "@foxglove/studio-base/reducers/layoutHistory";
 import mosaic from "@foxglove/studio-base/reducers/mosaic";
-import panels, {
-  PanelsState,
-  getInitialPersistedStateAndMaybeUpdateLocalStorageAndURL,
-} from "@foxglove/studio-base/reducers/panels";
 import tests from "@foxglove/studio-base/reducers/tests";
 import userNodes, { UserNodeDiagnostics } from "@foxglove/studio-base/reducers/userNodes";
-import { Auth as AuthState } from "@foxglove/studio-base/types/Auth";
 import { HoverValue } from "@foxglove/studio-base/types/hoverValue";
 
 const getReducers = () => [
-  panels,
   mosaic,
   hoverValue,
   userNodes,
@@ -41,32 +35,22 @@ const getReducers = () => [
   ...(process.env.NODE_ENV === "test" ? [tests] : []),
 ];
 
-export type PersistedState = {
-  panels: PanelsState;
-  search?: string;
-};
-
 export type Dispatcher<A extends ActionTypes> = ThunkAction<void, State, undefined, A>;
 
 export type State = {
-  persistedState: PersistedState;
   mosaic: { mosaicId: string; selectedPanelIds: string[] };
-  auth: AuthState;
   hoverValue?: HoverValue;
   userNodes: { userNodeDiagnostics: UserNodeDiagnostics; rosLib: string };
   layoutHistory: LayoutHistory;
 };
 
-export default function createRootReducer(args?: { testAuth?: any }): Reducer<State, ActionTypes> {
-  const persistedState = getInitialPersistedStateAndMaybeUpdateLocalStorageAndURL();
+export default function createRootReducer(): Reducer<State, ActionTypes> {
   const initialState: State = {
-    persistedState,
     mosaic: {
       // We use the same mosaicId for all mosaics to support dragging and dropping between them
       mosaicId: uuidv4(),
       selectedPanelIds: [],
     },
-    auth: Object.freeze(args?.testAuth || { username: undefined }),
     hoverValue: undefined,
     userNodes: {
       userNodeDiagnostics: {
@@ -78,15 +62,9 @@ export default function createRootReducer(args?: { testAuth?: any }): Reducer<St
     layoutHistory: initialLayoutHistoryState,
   };
   return (state: State | undefined, action: ActionTypes): State => {
-    const oldPersistedState: PersistedState | undefined = state?.persistedState;
-    const reducers: Array<(arg0: State, arg1: ActionTypes, arg2?: PersistedState) => State> =
-      getReducers() as any;
-    return reducers.reduce(
-      (builtState, reducer) => reducer(builtState, action, oldPersistedState),
-      {
-        ...initialState,
-        ...state,
-      },
-    );
+    return getReducers().reduce((builtState, reducer) => reducer(builtState, action), {
+      ...initialState,
+      ...state,
+    });
   };
 }
